Allow config to require other files.
This is useful for reusability of configuration snippets.

Required files are automatically watched for changes, just like the main
.phoenix.js config file.

Usage:

```javascript
require('<path to some.js file>');
```

If the path is a relative path it's resolved relative to the (real)
location of the .phoenix.js file. If the .phoenix.js file is a symlink
the symlink is resolved before resolving the location of the required
file.